watch-for-session-idle: Add a python variant of watch-for-session-idle.sh
In RHEL 8, busctl doesn't support --json=short and it fails to handle
--match rules properly.

This causes a flood of events to get generated making dbus-daemon
balloon to gigabytes of memory and makes the script wake up 100 times
a second.

So as an alternative to busctl, this commit adds a dbus-python
version of the script.