shell-lg: Load history at start up
This adds a jq dep.

There are a few other random clean ups and robustness
fixes.

I'm dismayed to find out `read -e` can't be given
a dynamic tab completion list, so I may just rewrite this
thing in javascript or something.

The whole point of it being in bash was so I could get readline for
free.