shell-lg: Improve inspect() feature
Right the inspector leaves an ugly box around what's picked.
This isn't so useful and also is sort of hard to get rid of
(since we can't just dismiss it by sliding lookingGlass up).

This commit gets rid of it immediately after picking finishes,
and also fixes up the code to handle escape better.