at-spi-snoop.py: Add start of a program to snoop events
Eventually I'd like to record everything i'm doing and feed it to
AI to automate writing status reports.  I spent an hour this morning
trying to get the goo outputed to json only semi-successfully.

I'm dumping it here for now, so I don't lose it in case I pick it up
later.