find-package-dependents.py: Add script to query dnf reverse dependencies
In the olden days we had `repoquery whatrequires --alldeps` but
that broke when `dnf repoquery` replaced it. There's a `--recursive`
but it doesn't seem to work, or I'm not using it correctly or
something.

This commit adds a python wrapper around `dnf repoquery --whatdepends`
to get the closed list of transitive reverse dependencies.